3th commit

+++ PageRankReducer.java
@@ -8,7 +8,7 @@ extends Reducer<Text, Text, Text, Text> {
 @Override
 public void reduce(Text key, Iterable<Text> values, Context context)
 throws IOException, InterruptedException {
-int PR = 0; 
+double PR = 0.0; 
 String val = null; 
 for (Text value : values) {
 String temp = value.toString();
@@ -17,7 +17,7 @@ if (temp.charAt(temp.length() - 1) < '0' || temp.charAt(temp.length() - 1) > '9'
 	continue;
 };
 String[] strs = temp.split(" ");
-PR += Integer.parseInt(strs[1]);
+PR += Double.parseDouble(strs[1]);
 }
 context.write(key, new Text(val + " " + PR));
 }
